Fix download mojo property 'total' type
Download mojo returns property 'total' (size of downloading file in
bytes). It has int32 type and in case there are file with size more
than aproximatly 2.14Gb it returns wrong value (int32 overflow).
This fix change type of property 'total' from int32 to int64.

Bug: None
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4833639
Commit-Queue: Alexander Yashkin <a-v-y@yandex-team.ru>
Cr-Commit-Position: refs/heads/main@{#1192929}

+++ chrome/browser/ui/webui/downloads/downloads_list_tracker.cc
@@ -284,7 +284,7 @@ downloads::mojom::DataPtr DownloadsListTracker::CreateDownloadData(
     file_value->url = absl::make_optional<GURL>(download_item->GetURL());
   }
   file_value->display_url = GetFormattedDisplayUrl(download_item->GetURL());
-  file_value->total = static_cast<int>(download_item->GetTotalBytes());
+  file_value->total = download_item->GetTotalBytes();
   file_value->file_externally_removed =
       download_item->GetFileExternallyRemoved();
   file_value->resume = download_item->CanResume();
